third_party/WebKit/LayoutTests/screen_orientation/screenorientation-api.html

<!DOCTYPE html>
<html>
<body>
<script src="../resources/js-test.js"></script>
<script>
description("Validates the Screen Orientation API is exposed.");

shouldBeDefined("screen.lockOrientation");
shouldBeDefined("screen.unlockOrientation");
shouldBeDefined("screen.orientation");
shouldBeDefined("screen.onorientationchange");
// Screen should be an EventTarget
shouldBe("screen.__proto__.__proto__", "EventTarget.prototype");
shouldBeDefined("screen.addEventListener");
shouldBeDefined("screen.removeEventListener");
shouldBeDefined("screen.dispatchEvent");
</script>
</body>
</html>
